Organize reder-list module by adding index
I also updated the adding task logic
