+++ src/index.html
@@ -21,8 +21,6 @@
                 </form>
             </li>
             <li id="todo-box">
-                <ul class="list">
-                </ul>
             </li>
             <li class="clear">
                 <button type="button" class="clear-completed">Clear all completed</button>
